requirements.txt
fastmcp>=2.10.4
python-telegram-bot>=21.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
pydantic>=2.5.0
httpx>=0.25.0
uvicorn>=0.24.0 